deployment/ECS-ResNet/requirements.txt
flask==3.1.0
boto3==1.35.43
numpy==2.0.2
gunicorn==23.0.0
tensorflow==2.18.0
pillow==10.4.0
python-dotenv==1.0.1
flask_cors==5.0.1
